+++ src/Graphics/Model/mdlFile.c
@@ -1387,31 +1387,31 @@ extern u32 DAT_0069cd00;
 extern u32 DAT_0069cd04;
 extern MdlStridedValue DAT_0069cd08[];
 extern u32 DAT_007cad6c;
-extern u32 DAT_007cad7c;
+extern f32 DAT_007cad7c;
 #pragma alias DAT_007cad7c_f32 DAT_007cad7c
 extern f32 DAT_007cad7c_f32;
-extern u32 DAT_007cada0;
-extern u32 DAT_007cadb0;
+extern f32 DAT_007cada0;
+extern f32 DAT_007cadb0;
 #pragma alias DAT_007cadb0_f32 DAT_007cadb0
 extern f32 DAT_007cadb0_f32;
-extern u32 DAT_007cadc0;
+extern f32 DAT_007cadc0;
 extern u32 DAT_007cadc4;
 #pragma alias DAT_007cadc4_f32 DAT_007cadc4
 extern f32 DAT_007cadc4_f32;
-extern u32 DAT_007cadd0;
+extern f32 DAT_007cadd0;
 #pragma alias DAT_007cadd0_f32 DAT_007cadd0
 extern f32 DAT_007cadd0_f32;
 extern u32 DAT_007cade4;
 #pragma alias DAT_007cade4_f32 DAT_007cade4
 extern f32 DAT_007cade4_f32;
-extern u32 DAT_007cae00;
+extern f32 DAT_007cae00;
 extern u32 DAT_007cae18;
 #pragma alias DAT_007cae18_f32 DAT_007cae18
 extern f32 DAT_007cae18_f32;
 extern u32 DAT_007cae4c;
 #pragma alias DAT_007cae4c_f32 DAT_007cae4c
 extern f32 DAT_007cae4c_f32;
-extern u32 DAT_007cae58;
+extern f32 DAT_007cae58;
 #pragma alias DAT_007cae58_f32 DAT_007cae58
 extern f32 DAT_007cae58_f32;
 extern u32 DAT_007cae98;
@@ -1423,20 +1423,20 @@ extern f32 DAT_007cada0_f32;
 extern f32 DAT_007cae90_f32;
 #pragma alias DAT_007caea4_f32 DAT_007caea4
 extern f32 DAT_007caea4_f32;
-extern u32 DAT_007cae50;
+extern f32 DAT_007cae50;
 #pragma alias DAT_007cae50_f32 DAT_007cae50
 extern f32 DAT_007cae50_f32;
-extern u32 DAT_007cae58;
+extern f32 DAT_007cae58;
 extern u32 DAT_007cae5c;
 extern u32 DAT_007cae60;
-extern u32 DAT_007cae64;
-extern u32 DAT_007cae68;
-extern u32 DAT_007cae6c;
-extern u32 DAT_007cae70;
-extern u32 DAT_007cae74;
-extern u32 DAT_007cae78;
-extern u32 DAT_007cae7c;
-extern u32 DAT_007cae80;
+extern f32 DAT_007cae64;
+extern f32 DAT_007cae68;
+extern f32 DAT_007cae6c;
+extern f32 DAT_007cae70;
+extern f32 DAT_007cae74;
+extern f32 DAT_007cae78;
+extern f32 DAT_007cae7c;
+extern f32 DAT_007cae80;
 extern u32 DAT_007cae90;
 extern u32 DAT_007cae94;
 extern u32 DAT_007cae98;
